Add r/w bit input to i2c master module
Originally, I was thinking of making it determine to go to either
READ_DATA or WRITE_DATA from the LSB of address byte (address[0]) but for readability
and consistancy I decided to make it a separate input and make address 7
bits.

Also I still need to properly test READ_DATA state since I need a proper
way to simulate a slave device on the testbench.